refactor: Split endpoint to transport and instance
Changed the endpoints code to rename the existing endpoint types to
"endpoint transport" and add the concept of "endpoint instances". A
transport is the method by which data is sent, while instances allow
describing multiple endpoints that use the same transport (e.g.
bluetooth profiles).

Also added new APIs to get the total number of possible endpoint
instances and assign each instance a unique index, which can be used
for tracking separate state for each endpoint in other code files.